Add LED Support to the board initialization.
Compile local sources before the ASF sources.

Remove the references to __ram_end__ from the _sbrk call. This was
causing the call to setbuf(stdout, NULL) to fail, raising a forced hard
fault. Unfortunately, neither this behaviour or the original behaviour
is correct. __ram_end__ == the end of the contiguous ram range, _end is
defined after the stack. _end > __ram_end__ is always true, meaning this
call will always fail.